+++ lib/src/utils/types.dart
@@ -1,0 +1,8 @@
+part of http_api;
+
+typedef VoidFunction = void Function();
+typedef NextFunction = Future<ApiResponse> Function(ApiRequest request);
+typedef NextHandler = Future<ApiResponse> Function(
+  ApiRequest request,
+  NextFunction next,
+);