+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.7.2] - 09 September 2020.
+- Added `ApiLink.next` constructor to allows in-place link creation.
+- ⚠️ `ApiLink` now cannot be used as mixin. 
+- Added `ApiLink.next` example to the readme.
 ## [0.7.1] - 07 September 2020.
 - Removed flutter from dependency (add crossplatform support).
   - Replaced Flutter `Key` class with `CacheKey`.